break out outputs into separate file since it's shared
