+++ aws/lambda-zip/deployer.go
@@ -4,29 +4,10 @@ import (
 	"context"
 	"fmt"
 	"github.com/nullstone-io/deployment-sdk/app"
-	nsaws "github.com/nullstone-io/deployment-sdk/aws"
 	"github.com/nullstone-io/deployment-sdk/outputs"
 	"gopkg.in/nullstone-io/go-api-client.v0"
 	"log"
-	"strings"
 )
-
-const (
-	KeyTemplateAppVersion = "{{app-version}}"
-)
-
-type Outputs struct {
-	Region               string     `ns:"region"`
-	Deployer             nsaws.User `ns:"deployer"`
-	LambdaArn            string     `ns:"lambda_arn"`
-	LambdaName           string     `ns:"lambda_name"`
-	ArtifactsBucketName  string     `ns:"artifacts_bucket_name"`
-	ArtifactsKeyTemplate string     `ns:"artifacts_key_template"`
-}
-
-func (o Outputs) ArtifactsKey(appVersion string) string {
-	return strings.Replace(o.ArtifactsKeyTemplate, KeyTemplateAppVersion, appVersion, -1)
-}
 
 func NewDeployer(logger *log.Logger, nsConfig api.Config, appDetails app.Details) (app.Deployer, error) {
 	outs, err := outputs.Retrieve[Outputs](nsConfig, appDetails.Workspace)
